[ビルド成功] V3+V2.2ハイブリッド実装完了 - WPF依存関係修正・OSS標準サービス実装・DocOrganizer.exe 292MB生成成功

+++ src/DocOrganizer.Infrastructure/Services/V3/HeicConversionService.cs
@@ -41,7 +41,7 @@ namespace DocOrganizer.Infrastructure.Services.V3
                     
                     // JPEG設定
                     image.Format = MagickFormat.Jpeg;
-                    image.Quality = quality;
+                    image.Quality = (uint)quality;
                     
                     // 出力ディレクトリ確保
                     var outputDir = Path.GetDirectoryName(jpegOutputPath);
@@ -156,8 +156,8 @@ namespace DocOrganizer.Infrastructure.Services.V3
                         : (ushort)1;
                     
                     return new HeicImageInfo(
-                        Width: image.Width,
-                        Height: image.Height,
+                        Width: (int)image.Width,
+                        Height: (int)image.Height,
                         FileSize: fileInfo.Length,
                         Format: "HEIC",
                         HasExifOrientation: hasExifOrientation,
